Add --flip-x argument to optionally mirror the output.
When used with the usual cookie cutter profile, this makes the input shape correspond to the shape of the final cookie rather than the shape of the cookie cutter sitting upside down on the printer.

This doesn't really belong in here (sweep should not know anything about cookie cutters, but just process the input files as they are), but it was the easiest point to insert it in the Inkscape integration.